requirements.txt
pywinauto==0.6.8
pywin32==311
gspread==6.1.2
google-auth==2.28.2
python-dotenv==1.0.1
loguru==0.7.2
requests==2.31.0
